requirements.txt
streamlit==1.28.0
mysql-connector-python==8.1.0
pandas==2.1.0